[DCGEO-189] Move bridge directory
