+++ apps/bfd-pipeline/bfd-pipeline-rda-bridge/src/main/resources/log4j.properties
