+++ apps/bfd-pipeline/bfd-pipeline-rda-bridge/src/test/resources/carrier.csv
@@ -3,3 +3,4 @@ INSERT|-1000001|-100000398|-100000398|F|O|71|28-Mar-1973|28-Mar-1973|30-Mar-1973
 INSERT|-1000001|-100000399|-100000399|F|O|71|31-Mar-1976|31-Mar-1976|02-Apr-1976|1|1|31143|1|1571.11|0.00|||A|1571.11|0|1571.11|0.00|64.16|1|PCP13750|J329|0|J329|0|||||||||||||||||||||||||1|PCP13750||||0|999269192|||||0|1|1|11|40|31-Mar-1976|31-Mar-1976|||||0.00|0|0.00|64.16| |0|0.00|1571.11|0.00|A|||15|||0|||0.0||||0
 INSERT|-1000001|-100000400|-100000400|F|O|71|04-Apr-1979|04-Apr-1979|06-Apr-1979|1|1|31143|1|889.66|0.00|||A|889.66|0|889.66|0.00|69.16|1|PCP13750|J329|0|J329|0|||||||||||||||||||||||||1|PCP13750||||0|999269192|||||0|1|1|11|40|04-Apr-1979|04-Apr-1979|||||0.00|0|0.00|69.16| |0|0.00|889.66|0.00|A|||15|||0|||0.0||||0
 INSERT|-1000001|-100000401|-100000401|F|O|71|20-Jan-1982|20-Jan-1982|22-Jan-1982|1|1|31143|1|1656.54|0.00|||A|1656.54|0|1656.54|0.00|64.16|1|PCP13750|J329|0|J329|0|E669|0|||||||||||||||||||||||1|PCP13750||||0|999269192|||||0|2|1|11|40|20-Jan-1982|20-Jan-1982|||||0.00|0|0.00|64.16| |0|0.00|1656.54|0.00|A|||15|||0|||0.0||||0
+INSERT|-1000001|-100000401|-100000401|F|O|71|20-Jan-1982|20-Jan-1982|22-Jan-1982|1|1|31143|1|1333.21|0.00|||A|1333.21|0|1333.21|0.00|54.16|1|PCP13750|P292|0|I10|0||0|||||||||||||||||||||||2|PCP13750||||0|999269192|||||0|2|1|11|40|20-Jan-1982|20-Jan-1982|||||0.00|0|0.00|64.16| |0|0.00|1333.21|0.00|A|||15|||0|||0.0||||0
