+++ apps/bfd-pipeline/bfd-pipeline-rda-bridge/src/test/resources/expected/expected-fiss.ndjson
@@ -1,0 +1,20 @@
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000001","hicNo":"T01000004A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"143.35","principleDiag":"J329","npiNumber":"8886687804","mbi":"1S00E00AA04","medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000002","hicNo":"T01000004A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"15167.87","admDiagCode":"C188","principleDiag":"C188","npiNumber":"8886687887","mbi":"1S00E00AA04","medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000003","hicNo":"T01000004A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"1819.15","admDiagCode":"R52","principleDiag":"R52","npiNumber":"8886687887","mbi":"1S00E00AA04","medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000004","hicNo":"T01000004A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"16638.58","principleDiag":"J329","npiNumber":"8886687804","mbi":"1S00E00AA04","medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000005","hicNo":"T01000004A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"86.81","principleDiag":"J329","npiNumber":"8886687804","mbi":"1S00E00AA04","medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000006","hicNo":"T01000001A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"86.81","principleDiag":"J329","npiNumber":"8886688182","mbi":"1S00E00AA01","medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000007","hicNo":"T01000001A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"85.13","principleDiag":"J209","npiNumber":"8886688182","mbi":"1S00E00AA01","medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000008","hicNo":"T01000001A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"77.49","principleDiag":"J209","npiNumber":"8886688182","mbi":"1S00E00AA01","medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000009","hicNo":"T01000003A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"2195.76","principleDiag":"J329","npiNumber":"8886614550","mbi":"1S00E00AA03","medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000010","hicNo":"T01000003A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"2195.76","principleDiag":"J329","npiNumber":"8886614550","mbi":"1S00E00AA03","medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000011","hicNo":"T01000003A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"2195.76","principleDiag":"J329","npiNumber":"8886614550","mbi":"1S00E00AA03","medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000012","hicNo":"T01000003A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"2195.76","principleDiag":"J329","npiNumber":"8886614550","mbi":"1S00E00AA03","medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000013","hicNo":"T01000000A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"23895.06","principleDiag":"D4959","npiNumber":"8888883658","mbi":"1S00E00AA00","medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000014","hicNo":"T01000000A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"23895.06","principleDiag":"D4959","npiNumber":"8888883658","mbi":"1S00E00AA00","medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000015","hicNo":"T01000000A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"23895.06","principleDiag":"D4959","npiNumber":"8888883658","mbi":"1S00E00AA00","medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000016","hicNo":"T01000000A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"23895.06","principleDiag":"D4959","npiNumber":"8888883658","mbi":"1S00E00AA00","medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000017","hicNo":"T01000001A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"22166.75","admDiagCode":"J329","principleDiag":"J329","npiNumber":"","mbi":"1S00E00AA01","fissProcCodes":[{"procCd":"5A12012","procDt":"2017-01-18"},{"procCd":"5A1C00Z","procDt":"2017-01-18"},{"procCd":"5A1213Z","procDt":"2017-01-18"},{"procCd":"5A1945Z","procDt":"2017-01-18"},{"procCd":"5A0955Z","procDt":"2017-01-18"},{"procCd":"5A02110","procDt":"2017-01-18"},{"procCd":"5A02215","procDt":"2017-01-18"},{"procCd":"5A05121","procDt":"2017-01-18"},{"procCd":"5A0935B","procDt":"2017-01-18"}],"medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000018","hicNo":"T01000004A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"25843.73","admDiagCode":"J329","principleDiag":"J329","npiNumber":"","mbi":"1S00E00AA04","fissProcCodes":[{"procCd":"5A0522C","procDt":"2013-10-05"},{"procCd":"5A02216","procDt":"2013-10-05"},{"procCd":"5A0955B","procDt":"2013-10-05"},{"procCd":"5A1935Z","procDt":"2013-10-05"},{"procCd":"5A05121","procDt":"2013-10-05"},{"procCd":"5A09458","procDt":"2013-10-05"},{"procCd":"5A12012","procDt":"2013-10-05"},{"procCd":"5A02210","procDt":"2013-10-05"},{"procCd":"5A1935Z","procDt":"2013-10-05"},{"procCd":"5A1C00Z","procDt":"2013-10-05"}],"medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000019","hicNo":"T01000004A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"27267.14","admDiagCode":"J329","principleDiag":"J329","npiNumber":"","mbi":"1S00E00AA04","fissProcCodes":[{"procCd":"5A02210","procDt":"2017-10-17"},{"procCd":"5A12012","procDt":"2017-10-17"},{"procCd":"5A0221D","procDt":"2017-10-17"},{"procCd":"5A1945Z","procDt":"2017-10-17"},{"procCd":"5A1945Z","procDt":"2017-10-17"},{"procCd":"5A05221","procDt":"2017-10-17"},{"procCd":"5A09458","procDt":"2017-10-17"},{"procCd":"5A05121","procDt":"2017-10-17"},{"procCd":"5A19054","procDt":"2017-10-17"},{"procCd":"5A0945B","procDt":"2017-10-17"}],"medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"dcn":"*0000000000000000000020","hicNo":"T01000008A","currStatusEnum":"CLAIM_STATUS_ROUTING","totalChargeAmount":"14149.36","admDiagCode":"M810","principleDiag":"M810","npiNumber":"","mbi":"1S00E00AA08","fissProcCodes":[{"procCd":"5A09558","procDt":"2020-07-27"},{"procCd":"5A1C60Z","procDt":"2020-07-27"},{"procCd":"5A1955Z","procDt":"2020-07-27"},{"procCd":"5A0955Z","procDt":"2020-07-27"},{"procCd":"5A1223Z","procDt":"2020-07-27"}],"medaProvId":"","currLoc1Unrecognized":"?","currLoc2Unrecognized":"?"}}
