+++ apps/bfd-pipeline/bfd-pipeline-rda-bridge/src/test/resources/expected/expected-mcs.ndjson
@@ -1,0 +1,4 @@
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"idrClmHdIcn":"*00000000000003","idrStatusCodeEnum":"STATUS_CODE_ACTIVE_A","idrBillProvNpi":"","idrBillProvNum":"","idrTotBilledAmt":"889.66","idrClaimMbi":"1S00E00AA01","mcsDiagnosisCodes":[{"idrClmHdIcn":"*00000000000003","idrDiagIcdTypeEnum":"DIAGNOSIS_ICD_TYPE_ICD9","idrDiagCode":"J329"}],"mcsDetails":[{"idrDtlToDate":"1973-03-28","idrProcCode":"00000"}],"idrClaimTypeUnrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"idrClmHdIcn":"*00000000000004","idrStatusCodeEnum":"STATUS_CODE_ACTIVE_A","idrBillProvNpi":"","idrBillProvNum":"","idrTotBilledAmt":"1571.11","idrClaimMbi":"1S00E00AA01","mcsDiagnosisCodes":[{"idrClmHdIcn":"*00000000000004","idrDiagIcdTypeEnum":"DIAGNOSIS_ICD_TYPE_ICD9","idrDiagCode":"J329"}],"mcsDetails":[{"idrDtlToDate":"1976-03-31","idrProcCode":"00000"}],"idrClaimTypeUnrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"idrClmHdIcn":"*00000000000005","idrStatusCodeEnum":"STATUS_CODE_ACTIVE_A","idrBillProvNpi":"","idrBillProvNum":"","idrTotBilledAmt":"889.66","idrClaimMbi":"1S00E00AA01","mcsDiagnosisCodes":[{"idrClmHdIcn":"*00000000000005","idrDiagIcdTypeEnum":"DIAGNOSIS_ICD_TYPE_ICD9","idrDiagCode":"J329"}],"mcsDetails":[{"idrDtlToDate":"1979-04-04","idrProcCode":"00000"}],"idrClaimTypeUnrecognized":"?"}}
+{"changeType":"CHANGE_TYPE_UPDATE","claim":{"idrClmHdIcn":"*00000000000006","idrStatusCodeEnum":"STATUS_CODE_ACTIVE_A","idrBillProvNpi":"","idrBillProvNum":"","idrTotBilledAmt":"1656.54","idrClaimMbi":"1S00E00AA01","mcsDiagnosisCodes":[{"idrClmHdIcn":"*00000000000006","idrDiagIcdTypeEnum":"DIAGNOSIS_ICD_TYPE_ICD9","idrDiagCode":"J329"},{"idrClmHdIcn":"*00000000000006","idrDiagIcdTypeEnum":"DIAGNOSIS_ICD_TYPE_ICD9","idrDiagCode":"E669"}],"mcsDetails":[{"idrDtlToDate":"1982-01-20","idrProcCode":"00000"}],"idrClaimTypeUnrecognized":"?"}}
