+++ apps/bfd-pipeline/bfd-pipeline-rda-bridge/src/test/resources/outpatient.csv
