+++ apps/bfd-pipeline/bfd-pipeline-rda-bridge/src/test/resources/snf.csv
