--- apps/bfd-pipeline/bfd-pipeline-rif-to-rda/src/test/resources/expected/expected-mcs.ndjson
@@ -1,4 +1,0 @@
-{"changeType":"CHANGE_TYPE_UPDATE","claim":{"idrClmHdIcn":"*00000000000003","idrContrId":"inval","idrClaimTypeEnum":"CLAIM_TYPE_ADJUSTMENT","idrStatusCodeEnum":"STATUS_CODE_ACTIVE_A","idrBillProvNpi":"invalidinv","idrBillProvNum":"invali","idrTotBilledAmt":"889.66","idrClaimMbi":"1S00E00AA01","mcsDiagnosisCodes":[{"idrDiagIcdTypeEnum":"DIAGNOSIS_ICD_TYPE_ICD9","idrDiagCode":"J329"}],"mcsDetails":[{"idrDtlToDate":"28-Mar-1973","idrProcCode":""}]}}
-{"changeType":"CHANGE_TYPE_UPDATE","claim":{"idrClmHdIcn":"*00000000000004","idrContrId":"inval","idrClaimTypeEnum":"CLAIM_TYPE_ADJUSTMENT","idrStatusCodeEnum":"STATUS_CODE_ACTIVE_A","idrBillProvNpi":"invalidinv","idrBillProvNum":"invali","idrTotBilledAmt":"1571.11","idrClaimMbi":"1S00E00AA01","mcsDiagnosisCodes":[{"idrDiagIcdTypeEnum":"DIAGNOSIS_ICD_TYPE_ICD9","idrDiagCode":"J329"}],"mcsDetails":[{"idrDtlToDate":"31-Mar-1976","idrProcCode":""}]}}
-{"changeType":"CHANGE_TYPE_UPDATE","claim":{"idrClmHdIcn":"*00000000000005","idrContrId":"inval","idrClaimTypeEnum":"CLAIM_TYPE_ADJUSTMENT","idrStatusCodeEnum":"STATUS_CODE_ACTIVE_A","idrBillProvNpi":"invalidinv","idrBillProvNum":"invali","idrTotBilledAmt":"889.66","idrClaimMbi":"1S00E00AA01","mcsDiagnosisCodes":[{"idrDiagIcdTypeEnum":"DIAGNOSIS_ICD_TYPE_ICD9","idrDiagCode":"J329"}],"mcsDetails":[{"idrDtlToDate":"04-Apr-1979","idrProcCode":""}]}}
-{"changeType":"CHANGE_TYPE_UPDATE","claim":{"idrClmHdIcn":"*00000000000006","idrContrId":"inval","idrClaimTypeEnum":"CLAIM_TYPE_ADJUSTMENT","idrStatusCodeEnum":"STATUS_CODE_ACTIVE_A","idrBillProvNpi":"invalidinv","idrBillProvNum":"invali","idrTotBilledAmt":"1656.54","idrClaimMbi":"1S00E00AA01","mcsDiagnosisCodes":[{"idrDiagIcdTypeEnum":"DIAGNOSIS_ICD_TYPE_ICD9","idrDiagCode":"J329"},{"idrDiagIcdTypeEnum":"DIAGNOSIS_ICD_TYPE_ICD9","idrDiagCode":"E669"}],"mcsDetails":[{"idrDtlToDate":"20-Jan-1982","idrProcCode":""}]}}
